pr: ci improvements

+++ src/display_drv.c
@@ -20,15 +20,15 @@
 extern void display_gpio_init(void)
 {
     /* configure P0.0 as an output and turn off */
-    GP0DAT |= (1UL << (24 + CS_PIN_NO));
+    GP0DAT |= (1UL << (uint8_t)(24UL + CS_PIN_NO));
     display_cs_off();
 
     /* configure P0.2 as an output and turn off */
-    GP0DAT |= (1UL << (24 + RESET_PIN_NO));
+    GP0DAT |= (1UL << (uint8_t)(24UL + RESET_PIN_NO));
     display_reset_off();
 
     /* configure P0.4 as an output and turn off */
-    GP0DAT |= (1UL << (24 + DC_PIN_NO));
+    GP0DAT |= (1UL << (uint8_t)(24U + DC_PIN_NO));
     display_dc_off();
 }
 
@@ -37,7 +37,7 @@ extern void display_gpio_init(void)
  */
 extern void display_cs_on(void)
 {
-    GP0DAT |= (1UL << (16 + CS_PIN_NO));
+    GP0DAT |= (1UL << (uint8_t)(16UL + CS_PIN_NO));
 }
 
 /**
@@ -45,7 +45,7 @@ extern void display_cs_on(void)
  */
 extern void display_cs_off(void)
 {
-    GP0DAT &= ~(1UL << (16 + CS_PIN_NO));
+    GP0DAT &= ~(1UL << (uint8_t)(16UL + CS_PIN_NO));
 }
 
 /**
@@ -53,7 +53,7 @@ extern void display_cs_off(void)
  */
 extern void display_reset_on(void)
 {
-    GP0DAT |= (1UL << (16 + RESET_PIN_NO));
+    GP0DAT |= (1UL << (uint8_t)(16UL + RESET_PIN_NO));
 }
 
 /**
@@ -61,7 +61,7 @@ extern void display_reset_on(void)
  */
 extern void display_reset_off(void)
 {
-    GP0DAT &= ~(1UL << (16 + RESET_PIN_NO));
+    GP0DAT &= ~(1UL << (uint8_t)(16UL + RESET_PIN_NO));
 }
 
 /**
@@ -69,7 +69,7 @@ extern void display_reset_off(void)
  */
 extern void display_dc_on(void)
 {
-    GP0DAT |= (1UL << (16 + DC_PIN_NO));
+    GP0DAT |= (1UL << (uint8_t)(16UL + DC_PIN_NO));
 }
 
 /**
@@ -77,6 +77,6 @@ extern void display_dc_on(void)
  */
 extern void display_dc_off(void)
 {
-    GP0DAT &= ~(1UL << (16 + DC_PIN_NO));
+    GP0DAT &= ~(1UL << (uint8_t)(16UL + DC_PIN_NO));
 }
 /*** end of file ***/
